cloud_functions/bigquery/requirements.txt
# Function dependencies, for example:
# package>=version

Flask
flask-cors
redis
google-api-python-client
google-cloud-bigquery